récupération user DONE et go to lobby after

+++ app/src/main/java/com/example/spacedimvisuel/api/User.kt
@@ -1,15 +1,14 @@
 package com.example.spacedimvisuel.api
 
 import android.os.Parcelable
-import com.squareup.moshi.JsonClass
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class User (
-    val id: Long,
+data class User(
+    val id: Int,
     val name: String,
     val avatar: String,
-    val score: Long,
+    val score: Int,
     val state: State = State.OVER
 ) :Parcelable
